+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 ## [Ideas/Upcoming]
 
 ## [Unreleased]
+### Added
+* DragMoveBehavior to allow a window drag move when hold down any UI element
 
 ## [1.0.0] - 2021-07-20
 ### Added